Handle preempted instances
When the slurmd is stopped, a script checks the instance's status. If it
is stopping, then it will mark the node as down. If the node is a
preemptible instance and wasn't suspended by Slurm, the node will be marked
as power_down so that Slurm will suspend the node and delete the node in
GCP.